# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Non-Blocking API - User Service",
    "version": "1.0.0",
    "description": "OpenAPI spec for /api/user/{userId}. Generated from current Zod schemas and route behavior (including error responses)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "paths": {
    "/api/user/{userId}": {
      "get": {
        "summary": "Get user with posts",
        "description": "Get a user and their posts from JSONPlaceholder, with SWR caching",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "User ID (positive integer)",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with user and posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid User ID parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TooManyRequestsError"
                }
              }
            }
          },
          "500": {
            "description": "Failed to fetch user data (upstream failure or unexpected error)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericError"
                }
              }
            }
          },
          "502": {
            "description": "Failed to validate user data (schema validation)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorWithDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "id",
          "name",
          "username",
          "email",
          "address",
          "phone",
          "website",
          "company"
        ],
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Leanne Graham"
          },
          "username": {
            "type": "string",
            "example": "Bret"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "Sincere@april.biz"
          },
          "address": {
            "type": "object",
            "required": [
              "street",
              "suite",
              "city",
              "zipcode"
            ],
            "properties": {
              "street": {
                "type": "string",
                "example": "Kulas Light"
              },
              "suite": {
                "type": "string",
                "example": "Apt. 556"
              },
              "city": {
                "type": "string",
                "example": "Gwenborough"
              },
              "zipcode": {
                "type": "string",
                "example": "92998-3874"
              }
            }
          },
          "phone": {
            "type": "string",
            "example": "1-770-736-8031 x56442"
          },
          "website": {
            "type": "string",
            "example": "hildegard.org"
          },
          "company": {
            "type": "object",
            "required": [
              "name",
              "catchPhrase",
              "bs"
            ],
            "properties": {
              "name": {
                "type": "string",
                "example": "Romaguera-Crona"
              },
              "catchPhrase": {
                "type": "string",
                "example": "Multi-layered client-server neural-net"
              },
              "bs": {
                "type": "string",
                "example": "harness real-time e-markets"
              }
            }
          }
        }
      },
      "Post": {
        "type": "object",
        "required": [
          "userId",
          "id",
          "title",
          "body"
        ],
        "properties": {
          "userId": {
            "type": "number",
            "example": 1
          },
          "id": {
            "type": "number",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "sunt aut facere repellat"
          },
          "body": {
            "type": "string",
            "example": "quia et suscipit suscipit recusandae"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "user",
          "posts"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          }
        }
      },
      "ErrorWithDetails": {
        "type": "object",
        "required": [
          "error",
          "details"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid User ID parameter"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "userId": [
                "Expected number, received string"
              ]
            }
          }
        }
      },
      "GenericError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Failed to fetch user data"
          }
        }
      },
      "TooManyRequestsError": {
        "type": "object",
        "required": [
          "statusCode",
          "error",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 429
          },
          "error": {
            "type": "string",
            "example": "Too Many Requests"
          },
          "message": {
            "type": "string",
            "example": "Rate limit exceeded, retry later"
          }
        }
      }
    }
  }
}